+++ src/components/inputs/radio/Radio.stories.customLabel.tsx
@@ -9,7 +9,7 @@ import { Radio } from './Radio';
 import { Text } from '../../basic/text/Text';
 import { Padding } from '../../layout/Padding';
 import { Row } from '../../layout/Row';
-import { Select } from '../Select';
+import { Select } from '../select/Select';
 
 export const RadioCustomLabel = (): React.JSX.Element => (
 	<Radio
