+++ src/components/inputs/select/Select.stories.multipleSelection.tsx
@@ -8,7 +8,7 @@ import React, { useState, useCallback } from 'react';
 
 import { Text } from '../../basic/text/Text';
 import { Container } from '../../layout/container/Container';
-import { Select, type SelectItem } from '../Select';
+import { Select, type SelectItem } from './Select';
 
 const items: SelectItem[] = [
 	{ label: 'Hi', value: '1' },
