+++ src/components/inputs/select/Select.tsx
@@ -9,16 +9,16 @@ import React, { useState, useMemo, useCallback, useEffect } from 'react';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
-import { getColor } from '../../theme/theme-utils';
-import { Icon } from '../basic/icon/Icon';
-import { Text } from '../basic/text/Text';
-import { INPUT_BACKGROUND_COLOR, INPUT_DIVIDER_COLOR } from '../constants';
-import type { DropdownItem, DropdownProps } from '../display/dropdown/Dropdown';
-import { Dropdown } from '../display/dropdown/Dropdown';
-import { Container } from '../layout/container/Container';
-import { Divider } from '../layout/divider/Divider';
-import { Padding } from '../layout/Padding';
-import { Row } from '../layout/Row';
+import { getColor } from '../../../theme/theme-utils';
+import { Icon } from '../../basic/icon/Icon';
+import { Text } from '../../basic/text/Text';
+import { INPUT_BACKGROUND_COLOR, INPUT_DIVIDER_COLOR } from '../../constants';
+import type { DropdownItem, DropdownProps } from '../../display/dropdown/Dropdown';
+import { Dropdown } from '../../display/dropdown/Dropdown';
+import { Container } from '../../layout/container/Container';
+import { Divider } from '../../layout/divider/Divider';
+import { Padding } from '../../layout/Padding';
+import { Row } from '../../layout/Row';
 
 const Label = styled(Text)<{ $selected: boolean }>`
 	position: absolute;
